Scrabble words/Grid print/tests setup
I added the list of scrabble words I found on the internet. I got grid working I believe, although I haven't done extensive testing. This included making location work with equals and hashcode for the filledSquares. The grid simply has a hashset of the filled squares and now has a bounding box of the size so that the O(N) operation of how much space has been taken up doesn't need to be performed. Locations and Grid need tests. Specifically to test the stuff I believe works now, I just want to make it nice and formal, and more extensive. Right now everything is just with print statements  and stuff because I didn't reconfigure the project with junit and such which I need to do to set up the tests. Overall though gettign reasonably near the mechanical portions of the game working.

Everything else!