Testing for Grid/bug fixed
Started creating the tests for grid. Noticed that the order of the array was causing errors, and since we don't care about the order I made it a HashSet instead.

Have to figure out the code coverage but I believe we have reasonable coverage for the grid and everything seems to be working well. Can probably start moving forward with the game elements now or soon