Built out player/made game
Player is just handling the aspects that a player MIGHT have to handle. The game will handle making all of the player's doing everything together, natually. So this class is reasonably simplistic. Player does not handle the idea of a "person" playing multiple games and thus has no awareness outside of the games they will be within. I gave tile hashcode so that when you dump a tile it can be found within the hand that is a HashSet. It's also just useful to have for these sorts of comparisions in general. Other than that