Adding general functions/misc
Bug: There was a bug in which the bag would have nulls within it rather than letters up to a certain point and the nulls as designed. Naturally this was an issue with removing from the bag.

Added toString to be completed to many of the classes, as well as adding GameTest and PlayerTest. the singular test that is being creating in GameTest is not working as intended and the behavior of Peel in regard to the tile bag is acting quite weird (my first guess is the issue with the bag initialized to each players). Added simple functions to game and changed players to a HashSet. I might change this back though because I don't ever really need to find a specific player and it takes a little more memory