Actually fixed tile bag bug
Tile bag removal had a bug in which removing one or two elements worked but after that would would start to be nulls within the array. This was a simple fix (and silly error on my part) and was a result of grabbing a random element from the entire array instead of from the available elements. So, after removing one or two elements a null element would get randomly selected because the bag wasn't full.