Running small game test
In the process of creating the small game test I realized a few things. First, I forgot to add a remove tile capability for the grid and the player, which I quickly added for the test to have player 2 "accidentally' misspell a word, as might happen in an actual game. Similarly, I realized that Grid should have the getTile utility function to get a tile a given location instead of just locationFilled. I renamed gridDone() for player for the test to canPeel() because this makes a lot more intuitive sense and this is how it will be used in the context of the game (the player will be able to peel if teh canPeel() check is passed). Considering having player peel() have canPeel within it to make sure that the player cannot somehow accidentally trigger a peel (even though on the front end the button or activation mechanism would be disabled if canPeel is false) it might be a nice saftey mechanism. Otherwise, I guess it is technically superfluous so I  could always create a SafePlayer class that extends it if I really want to. Fixed a small java syntax silly mistake in GridTest in which I wasn't using characters properly, otherwise not important to this push. All of these changes and functionality shifts were in favor of making the game within the small game test run more smootly.