More GameTest + BUG
Added another test for the game to see if game complete worked in the circumstance in which a peel can still be performed, (i.e enough tiles to give out to the players). It did but I made the hand a HashSet and didn't think about having multiple identical Tiles in the hand. To solve this bug I switched to a HashMap for the hand of the player where it would keep track of each type of tile. However, this added some wrinkles to adding and removing tiles from the hand in those spots would could probably be encapsulated in a MultiSet class. The bug is fixed but in the tests for the grid .getHand().size() since we are using a HashMap that stores how many of each tile there is this only gives the number of keys, or unique tiles. This happens to end up working for the tests because all of the tiles are unique but the method size within game should really get the number of tiles within the hand. When checking if a tile is contained within the hand using the keySet from the new HashMap works well and poses no issues, so while a slight change nothing else needs to be changed.

So, while it works now I'm gonna make the HashMap keeping track of each type of how many of each tile it's own class to clean up the add and remove methods within the Player that are now a bit muddy dealing with the edge cases and underlying logic of keeping track of how many of each tile (removing the key if only one left, incrementing and decrementing, etc.). Additionally, this class could handle the issue with size.