Fixed player hand as set
I fixed the issue of the player's hand being represented in a set. that is used within the Player class naturally, a player can have multiple 'a' tiles, 'b' tiles, etc. so a set doesn't work. Simple oversight on my part. To remedy this the MultiSet data structure was made and used for the hand to have the same quick functionality for finding an element, adding an element, and removing an element, as well as the number of elements within the set. This data structure allows the tests to use set functions and consider the hand as a set (that of course can contain duplicates) with .add, .contains, .remove, and .size which importantly gets the number of elements including duplicates, resolving the issue in the testing for the last commit. Also, I renamed the tiles the player has but have not put on the Grid the players hand within the class because calling them the tiles was vague and confusing.