Setting up AI Objects
Setting up for AI I added the MinIndexPriority queue for A* that will be used, which was copied over from another implementation from "Algorithms, 4th edition by Robert Sedgewick and Kevin Wayne". Also began making the AI players and the branchable interface for the A* implementation that I will complete. For the branch method in AIPlayer to be complete we need to be able to copy a grid with a copy constructor to make the tree of seperate object, so that is still a stub, although much of it is complete. A* is very incomplete but the objects (implementation, extensions, etc.) should be set up for the constructor which will perform the search.