Reorganized tests to AI/game + others
I tweaked the wordSearchGrid. Now I realize all of the methods we would need to be good general purpose Grid methods, so for now all of that has been moved over and there is no special AI extension of Grid. Also, the AIPlayer will be branching Grids and their hand is needed to branch, so that no longer extends Branchable.

Branchable is fixed so the branch function must now return the same type as the class

All the grid class variables are protected so those potential extending classes (again don't really do anything now and likely not in many of the future versions) can actually access these variables.

Changed the testing so now there is a AI testing and game testing folder within the larger testing folder. I created the AIPlayer test for future use but haven't put any tests in there as AIPlayer branch needs to be completed

Began on AIPlayer branch but it is still not complete. We will be branching into other valid boards with words that are all acceptable, so we need to branch into all other valid words with the tiles gives that are valid words. A small little algorithm I have to code up and test due to the edge cases and issues that might arise.