Working on AI Branch Algo
Since we are branching based on words we need an algorithm for all possible words at each location that can be placed based on the player hand and the current grid. This requires matching the specific substring within the word at the location either up or right but also requires a comparison of two MultiSets. So, the multiset was updated with additional functionality that supports these operations, subsets, addTo/union, a copy constructor, string to MultiSet, contains, etc. With this we could do the two portions of validating a word for A* search. My concern is that the branching factor is going to be too high later, but we will cross that bridge when we come to it.  AIPlayer branch is a complete mess right now as I was tinkering with other approached and helper functions which I have commented out. I think this approach will work reasonably.

-- Some of this functionality could potentially be made into a WordSet class (that could handle finding all words with specific substrings containing specific letters, likely with a tree structure that has characters along the path of the word for efficiency, as well as HashSet for quickly finding words, despite potential memory costs?). As a side note a Trie might be better than a HashMap in this application considering the longest string I have is not that long in the scrabble dictionary and the big O for finding the word is not that long.

Also, I made a data structures test folder with a file for testing the MultiSet. Nothing is in it though yet...