Working on word find + PROBLEM
I worked on place word which seems tricky and probably has some bugs. The issue with the algorithm is it doesn't consider the edge case in which the word goes through multiple disconnected letters (think of forming a box in the game, it only considers one side of the connection, while the other immediately disqualifies the word). So, while this should work to create an A* search that works pretty well, it does disqualify one of those edge cases, so I'll probably rework it. Bummer.