Branch forward/small accompanying changes
Created branch forward function for ONLY words down, hopefully fixing the mentioned problem in the previous commit.

Small Other Changes:

Adding methods to grid that were useful for the AI but would be handled by a grid object. Many of these functions are specifically useful in the context of AI or A* so I may want to make a Grid that extends the functionality of the Grid for an AI context and flesh out. This class already exists in the project, I just this temporarily for the sake of simplicity.

Made game protected within player so that the extending class AIPlayer could access it. This makes perfect sense I just hadn't done it, so no issues with visibility with this in the future.

Since Grid has two constructors that take in a single parameter the tests had to be updated so that null was "cast" to a HashSet<String> to specify which constructor needs to be used.

Next:
This is a sort of intermediary commit so I need to test that the down version of the letters is working properly, then add the right direction version in the AIPlayerTest which was created but currently has no tests.