Cleaned up branching algo tests
Formalized the print statements I was using to see if the grids were looking good and creating the words I had hoped. I did this with the MultiSet, but I needed to add a comparison method and HashCode associated with equals. I also updated the getWordsPlayed() method to return a multiset for the same word played twice on the board (this might have an EXTREMELY slight cost to checking if a board is valid but I really don't mind). I also deleted many of the extraneous print statements within the tests (except for printing the player's hand as it is unclear due to seed randomness)