Renamed methods to CamelCase/wrote branch backward
Wrote up the branch backward method which essentially searches for words and removes them, placing the tiles back into the hand of the player. It works by finding the start of a word in the specified direction and removing all of those letters as long as they aren't part of another word. Branch backwards has not been tested yet though, and the file to do so it set up but currently empty. The functions in AIPlayer were not in camel case for some reason which was incredible inconsistent so I fixed that as well.