Tested/Debugged Branch Backwards
Generally just debugged branch backward. Had some small issues, including having to check for the same grid being added twice, requiring me to implement HashCode for the grids (this also was possibly a small issue with branchForward I didn't catch because it only appeared in an edge case that wasn't really tested. Also for some reason I made the locations for branchBackwards/branchForward a function when it could've just been a HashSet cause the other stuff was a function. So I cleaned that up too.