Created A Star using HashSet without relaxation with small test
Created a star with HashSets, which isn't super effective but works along with some small test cases for it. The test case for an empty board DOES not work but only because of a bug in branch in which an empty board cannot branch because there are no words to branch off. This bug needs to be fixed, but when it is this should work. The tests are also very informal right now, just giving print statements, but they indicate that the proper path is being created.

The heuristic function is just how many tiles are in the players hand (essentially just removing the constraints of the game) while the cost from one AIPlayer copy to the next is the change in the number of letters, so removing letters to hand is a negative value while adding letters to the hand is a positive value.