Branching on empty board implemented
Basically a simpler version of the general case, we can now branch on an empty board. Created a test case confirming the overall branch method now works for the empty edge case. Added isEmpty() to grid because while technically we can figure that out having a method within the grid object itself for that is nice. Naturally I also added this new test to the AllBranchTests suite