Diagnosing AStar which is broken
The branch function for a blank grid works, however the AStar is extremely slow and doesn't really work. It works for when there are very few potential words or when the word is only one away, however aside from that it gets EXTREMLELY large, maxing out the current limit of size in the min index priority queue (which can be fixed, but not a priority at the moment). So, I threw together a really yucky Maze class that can be used to test A Star. Everything is public and it isn't great, but it does the job and can branch. Based on the maze example, it seems like A star is the culprit and not the heuristic functions, as well as the fact A star is exhibiting behavior that is shouldn't be due to implementation errors on my part. One issue with A star at the moment that is causing it to balloon in memory requirements is that it is collecting duplicate objects because we do not check if the OBJECT has already been visited, just the index which we always add a new one. Not using arrays and creating the tree as we go is a bit of a mess and this definitely needs to be reworked. The HashSets in this implementation were never supposed to be super memory efficient but this is just very broken. Now, maze should be a good tool to see how everything is working. At the moment, although it is hard to tell, it may not be exploring efficiently according to the heuristic as well. This may be I did not implement relax because I wanted a quick and dirty test.  The good news is that with infinite time and computing power it is finding the correct path in the maze and in bananagrams, eventually....