Debugging A*/A* Working Somewhat
Implemented HashCode for maze to get that working, but A* stll gets very large when the path is very short (for example, a very thin and tall maze where the goal. Rewriting it slightly and handling the objects properly in visited made it work for both the maze example and for bananagrams. HOWEVER, it only worked for some combinations of letters in bananagrams in a reasonable amount of time (usually common letters, like e, d, t, m, a, etc.) and couldn't handle j, w, etc. except for in simple situations I created. Additionally this was ONLY on the 1000 words file. Otherwise, it cannot handle the number of words very well and just sits and doesn't spit out a result. The issue with j, w, z, etc could potentially be fixed with a better heuristic function, probably inversely related to the number of words with that word could be the number of points for the heuristic. However, the A* implementation is extremely new and has unfinished issues despite it working for these simpler examples. It needs the relax portion done, and needs ot be able to drop the MinIndexPQ eventually. Most important though, it uses HashSets when an array could be used in some places, which is inefficient for memory. Changing this, we could hope for some improvement. There also might be some bugs that decrease efficiency lurking, but those are tricky to spot and I'll see how this helps first.

Also added some small debugging functions within A*