Created A Star version with array
Relax on the already visited case has not been created for either of the A* searches, but the array version appears to work much better. The issue seems to be in the speed of the branch algorithm. There a few ways I can think of to improve this (after implementing relax and making the A* properly done). Dumping also needs to be added within the branch function as well, and tests need to be added for the cases in which the player already has a mostly completed grid.

-- Improving the branch algorithm by considering words in parallel instead

-- potentially implementing a data set for the words to eliminate large swaths of words at once instead of going through each of the m. (26 HashSets for words containing 'a', 'b', 'c', etc. and only going through the sets for the letters in the hand of the player).

-- Making the branch algorithm terminate after 10 words or something small so that larger datasets of words like the scrabble set can be used without considering too many options at once, since the ultimate goal is really just to get any solution. An alternative to A* could also be created where when it find a word that is sufficiently good (based on a set value and heuristic function) and moves forward from there, not returning to branch more unless all of those options are exhausted. However, this augmented A* would be quite tricky.