A star with hashset/array implementation (no relax) (#5)
A star works for creating a board for smaller number of letters and a maze, and letters that are not difficult. Otherwise it takes a long time (likely because of how long it takes to branch). It doesn't have the "relax" for when shorter path is found to an already visited node. I will complete this soon but suspect it doesn't have a huge effect for this specific problem (obviously it is critical in many general A* applications). Some tests for this were created to test a small maze object for testing as well as empty/partially completed grids. It only starts to struggle in the bigger example, and cannot handle the entire scrabble dictionary due to the number of possible words. Since most of the words in the scrabble dictionary don't even work, I suspect it is a problem with branching and not A*, as I don't think the branching factor isn't a ton higher even with that many words. It's just processing them that is the issue.