Relax implemented for ArrayList/HashSet
Implemented relax for the ArrayList and HashSet implementation. In order to find the correct I need to add a HashMap to BOTH implementation to get the index from a given object so that we could determine. There is likely a better way but this allows the relax functionality to work, as far as I can tell. I tested this on the Maze because that has more important situations in which relax is called, but as far as I can tell is working. However, the tests are still very small and this hasn't been stress tested on a case were relaxing is critical. Also fixed the HashCode for the Maze object.