Set up multithreading for Branch
For the small grid stress test (using all of the scrabble words), multithreading improved the performance by a factor of five on my laptop. On a larger computer I suspect it would improve it by more because there are more threads to allocate. Potentially, a GPU could be used for this at some point but it would require an external library and extra things to be done, so if I do this for the branching algorithm I will probably do it far later. I modified the existing branch algorithm, but just to display how much better this variation works, it may be fruitful to have a Serial version in which the branch method has no parallelization. This is a HUGE optimization from before, but it's worth emphasizing there are other ways to improve performance as well significantly that have not been implemented (that I think of). Namely, creating hashsets for all of the words that contain a, b, c....z, and aa, ab, ac..... so that we can quickly grab from only those words rather than having to serach through the set of scrabble words that we have, and stopping the branch algorithm after N words are found because the branching factor is so high, the branch method is so costly, and there are so many valid solutions for most combinations of letters, a branch method optimized like this or optimized to find any solution rather than the best solution could be better. Maybe even something like if we get a word with a certain length (like 50% of the hand), stop immediately and use that word now. I'm sure there are others, but branch is where most of the computation time is being spent within A*.