A* with relax/HashSet and ArrayList versions (#6)
Implemented the A* using both HashSets and ArrayLists, including relax for both when a more efficient path is found. An extra HashSet did need to be added to both for this capability though. A* currently just has a max size of 100000, which could easily become a parameter or be implemented in a way such that the min index priority queue doesn't have a set size as was copied over from Algorithms, 4th edition by Robert Sedgewick and Kevin Wayne. Nearly all of the "new lines" are 11,000 new lines are from new word sets that are smaller than the scrabble word set (1000 words and 10000 words) being added for use and testing.