Parallelized Branch Algorithm (#7)
Took the previous branch algorithm and multithreaded it since comparing many of the words are small easily parallelized tasks. This greatly improved the performance on my laptop, and will likely improve the performance significantly on other devices as well (especially devices with more potential threads to use). The serial version of the branch is still within the project as AIPlayerSerial for performance comparison purposes, and just as a general alternative.