Reorganizing Player/WordSets
Made the Serial AI player and the Parallel AI player. The parallel one performs much better, but having for performance comparison both are nice. Since the WordsSet is going to have more capabilities to improve the performance of performance of branch within the AI Player in the future, I created a class for that instead of just using a HashSet. Especially since it is always a HashSet of strings and has a distinct purpose, I am making a WordsSet instead of having the AIWordsSet extend HashSet<String>. The "word search grid" or AI Grid which has been sitting around but not used could be used at some point to handle much of the branching methods within the player, aside from the parts involving the Hands of the players. (i.e, the grid given a hand could branch and return the hand that the player would have). Either way seems like a somewhat reasonable decision to me though.