Created AI Words Set that could be used to optimize AI, and appropriate test
The test is quite bare bones and doesn't have a lot of assertions. Essentially it just verifies by inspection that all words are placed properly into each of the bins for each letter. Hasn't been put into the Grid and AIPlayer yet though. Might need some reorganizing of the classes for that.