Grid Copy in branch was copying wordsSet
This is a kind of dumb bug, but the copy in the branch method was copying all of the words (all the scrabble words, 10000 words, or whatever) each time it branched rather than just using a shared word set. This caused an absolutely ridiculous amount bloat and now all of the scrabble words actually works. Sometimes it's quite slow with certain letter combinations anyway, but it actually work (generally) now. Literally a one line fix. We could still improve it potentially with greedy strategies and such, and will have to reevaluate how much parallelism helped.