Made AIPlayer interface/reworked tests
Took the various implementations of AI Players and made interfaces for them so that the stress tests can be more flexible and take in different types of players. For the smaller branch tests, since the AIPlayer class does not require anything but branch (no branchForward, branchBackward, etc.) I went with the AIPlayerParallel to test the functionality wihin the other tests. These tests are not stress tests so the AIPlayer doesn't matter while in the stress tests being able to compare is much more important. Will have to compare some of the changes on the experimental branch to see if the make significant improvments,