Peel test for externally putting letters in
I created a test that using the way the bag is set up allows you to make a grid with the AI and then "peel" a new specific tile from which the AI works. Makes it easier to test peels and things, but could probably be set up in a nicer way to be able to solve a real life game rather than simulating the tile bag in the code.