Cleaned up TODOs
Most of the TODOs were vestigial remnants of implemented features in which the comments were not removed. Many of them were speculative as to what should be changed in the future. I left three main ones around I should implement.

1. Creating a copyable interface for the classes than can be copied that implement the copy() method to make copyable more understandable

2. Implementing the begin for grid as well as new constructors for a combination of multiple types of players, rather than just specifying the number of players.

3. Altering the min index priority queue so that it dynamically grows rather than having a fixed size limit.

The last thing, while not in the current TODOs, is to integrate the A* into the player class, with a solve method and a solveAndUpdate method (or something similar) to solve and then update the grid accordingly.