Dynamically Growing IndexMinPQ for A*/A* abstract class
Created the dynamic min index priority queue that grows when at capacity to be used for A* in this case for both implementations of A*. I changed the private methods within IndexMinPQ to protected so that this extension could be done. The tests for A* were also reorganized so it is easier to set up examples based on a set or starting letters and the next peels. Because of the significant overlap between the implementations of A* (and the fact I essentially did the same thing to have both use this new dynamic IndexMinPQ) they should probably be one interface that each implementation implements instead. When comparing the results on the dybamic vs static sized queues there were comprable results. For the small example test I got 927, 912, and 874 for the static and 886, 905, and 899 for dynamic. For the larger example I got 14271, 14419 and 14214 for static, and 14419, 14592 and 14321 for the dynamic. Also created an abstract class for A* to handle the two implementations. Now, instead of A* occurring when the object is constructed it occurs when compute is run. After this is run the information about the A* can be gathered with the other functions outlines in the abstract class. There aren't any guardrails around using the information gathering methods before running the compute method, but the methods don't work if you try, it's just that no special error is thrown. In the tests for A*, the astar needed the .compute added so that the information from the  A* was actually present.