Broke up Player in helper classes
The methods with the Hand can probably just be encapsulated in Player, so that the visibility of Hand can be package and just held within the Player. Had some issues getting this working because I had to make suer all of the objects were passing all of the information to each other and specifically the Hand class was using the MultiSet rather than extending it. Player tests would make it easier to see the issues and test any future changes. Because of this new hand object some of the tests needed to have the object changed from MultiSet<Tile> to Hand. However, if Hand is fully encapsulated within Player the methods of player should be able to replace getting the Hand or MultiSet<Tile> entirely. As a side note, for some reason the endIndex was not set at the end of the A* implementation of AStarHashSets, so I fixed that.

Things to do:
1. Encapsulate Hand within Player so that player classes can be used within the test and hand can be used and accessed through the player.
2. Add the A* capabilities within the A* player so that instead of the A* class outside of the the player taking the player the AIPlayers have the compute method to get an A* result. Maybe some kind of interface could indicate this as well.
3. With the broken up player objects, create a "cheat player" that uses the GridArranger but it's own CheatBroker that instead of randomly drawing tiles can pick tiles that the real life player playing bananagrams can see and plug in to solve rather than the work around currently used some of the tests to demonstrate this capability.