Reorganized and Modularizing Player BUILD FAILS
Folders have been reorganized partially. In particular, I broke the player up into the portion of player that handles the Branching and the Astar. The overlapping logic within the very similar branching types was placed into the AbstractPlayerBranchingMethods which handles much of the overlap, and while hidden outside of the package, is an intermediatory to the Branch classes that complete the implementation details, both of which extend from this Methods class. Outside of that I split the player into a Broker and a GridArranger for modularity and for making the eventual CheatPlayer (for using the solver in a real game) smooth and clean to implement. The branching player naturally extends from player with the additional branching capability. The A* player from which the CheatPlayer will eventually extend needs the branching and the A* classes, each of which will have a corresponding factory which implement Buildable so that information from game that is used to create the objects isn't exposed to the outer scope. Now the branching methods use the new branch player while the astar is handled manually within those tests. More tests should be made to test the AIPlayer itself. Much of the parts work even though the build fails.