Built factories for AStar and BranchingPlayer
Got the factories working, although they have so much overlap that there should be probably be a factory they extend from (and each of these would be small classes extending that factory, or potentially just specific generics of the outer class as factories would live within the player). The generics used for everything need to be fixed up a bit (or at least checked to make sure that they are correct and not too loose). Tests (even though more for new stuff are needed) pass again although build still fails. These factories work for any class extending the abstract and are given parameters in the constructor for dependency injection. They use these parameters and the passed class with reflection to allow an object to be build with the give class and injected dependencies via the build method, which is enforced from the Buildable interface.