Setup AI Player w/o tests (Not cheat player)
The class organization is really weird now and again the generics between classes are a bit funky. A* should maybe be static, and updating the A* the AIPlayer is using by creating a new A* object with reflection in the factory makes some sense since A* probably shouldn't be able to have any parameters updated throughout. I'm not making it static (at least for now) because the current player (start parameter within the A*) can be updated and set now for the AIPlayer while the cost, heuristic, and isGoal are constant, so "steps" can be taken with the A* like in some of the tests.