Connected everything (doesn't work)
Cheat player is made, but still doesn't work in the small test I set up. Have a plethora of TODOs everywhere but now (except for the DefaultCheatPlayer) everything is connected properly and doesn't have building issues. Part of this was making the player use the general AbstractBroker (the object by default is still the HumanBroker right now, although I'm not sure if I like this design choice) so that the CheatPlayer can take it can change it into a CheatBroker which of course extends AbstractBroker. Generics still need fixing though. Since it is a pain to make the TileBag implement the collection interface and both the Queue for CheatBroker and TileBag must be acceptable for the getBag method, I just made the return type object which will have to change. This caused me to have to change it in player and add casts to it in the parallel and serial branching players. I should probably create a default player class as well that has these settings.