Fixing up generics slightly/"fixing" constructor issue
The constructor field was having issues within the A* implementations because of the generic type. For some reason making the abstract type didn't seem to work? There is almost certainly a more elegant solution but I just made the input to the constructor an object which is then cast (unchecked) into type T. This seemed to fix it. The issue is a cast to TileBag doesn't work now because the CheatBroker uses a Queue. So, next TileBag should be an implementation and there should be types of TileBagable, like CheatBag (the queue) and NormalBag. Then this issue wouldn't occur.