Kind of a disaster of objects...fixing
The reorganizations issues are becoming more appearent. Many of the implementations were handling the  broker, but now the broker is handled by the CheatPlayer, which had to be fixed. Also the type of bag and broker are very tightly coupled and things should be reorganized for that. Generally, while the flexibility of being able to pass the A* and branch that would be preferred, it would probably be better just to have an abstract class that can be extended to implement the A* and branching types. The game being null for many of these tests is a very annoying problem which I worked around by creating other constructors to take the bag directly in that case. However a dummy game object might be nice, but we will see after properly organizing all of the player types (along with the brokers, gridArrangers and game). The copy method wasn't working for the same reason with an unneeded with a depenedence on a game object. I'm just pushing this version to this branch for now to build from, hopefully.