Reconfigured player class heirarachy
After reconfiguring the player classes, we have it so that the abstract classes do not implement broker at all (for instance, the general player class no longer has "default" broker and gridArranger, and instead a NormalPlayer class inhereits from this. I realized I overthought the input on the player for the AIPlayer class as instead of creating a factory a player could be passed in and the information could be held in the player class (rather than passing information that would be needed and then having the factory build it). The only potential "issue" would be that a player with a grid that already exists could be given to the AIPlayer, but I don't really care as that's not really an issue, and if I did an extending class utilitzing the factory separately to use the constructor for this AIPlayer could be created instead rather than mashing too much functionality into one class. I've decided the "fancy" constructors that can take in a bunch of types and such and build the players/other stuff within the class should only be in the concrete implementations to simplify things and to make it easy to configure for those concrete classes (even if technically there would be a slight overlap across concrete classes implementing the same super class it would only be for trivial code). AI player should maybe be renamed becuase it isn't really a player in a sense because it doesn't implement Player, it just has a player within it. Possibly the CheatPlayer could use this factory so that the constructor could be called that creates a new player type within the class based on the type of player desired. However, this isn't really a huge concern either. Even though this kind of setup for the players may mean the broker object is exposed so that programmers can create their own broker objects with players, this isn't really a huge drawback and it prevents factory and reflection confusion. I also deleted the unnecessary folder. Overall, still really messy though. Tests do not work because the branching players do not support those constructors anymore (a small extension class can with a specific broker type)