CheatPlayer working for smalll test
Trying to deal with abstracting the shared parts of Serial and Parallel player but it's making the code really messy and is being a pain. I'll just go to the old version later and then that should fix if the player issues. Many of these thing with various implementations could possibly all have their own factories later. But before that the normal tests and cheat tests need to work, some more tests should probably be added, and the project needs to be more reasonably arranged.