Cheat Player Peel Examples
A bug where a blank board is valid if you have only one letter. This has been around a while and isn't really important since that wouldn't happen in a game but is an edge case to fix. There was a silly little bug in the CheatBroker class in which the copy was not working properly for the bag associated. This caused issues for the new tests that use peels because the "CheatBag, represented by a queue essentially, was not being copied over. This CheatPlayer seems to behave a little differently in word choice than the analogous AStarGrids test for reasons I am unsure of. It seems equally if not more efficient though. A new test is much slower with the new CheatPlayer for some odd reason. So while they both work there must be some issue with the ordering of peels or something. I noticed the using really large words makes it hard to peel and add another tile because removing that word means it must start all the way over.