Fixed up extra constructors for tests
Not sure if these constructors are useful or should stick around but I created them to test all of the tests. Now the build compiles and all of the tests pass (although many of them are print statments that need to be inspected). Importantly, the player and A* along with the CheatPlayer are working in the larger examples based on inspection so there shouldn't be any significant bugs in that regard. However, the files still need to be cleaned and reorganized. Also, another better heuristic function could be created. I found this article about bananagrams strategy, but in general I think placing hard letters first, creating a wide grid, and using medium length words (so that words can be removed on a peel without starting all the way over) could be good. Website: https://bananagrams.com/blogs/news/how-to-win-at-bananagrams